refactor: move DesignSpace to space.py
