+++ experiment_design/variable/__init__.py
@@ -1,8 +1,8 @@
-from experiment_design.variable.types import Variable, VariableCollection
+from experiment_design.variable.space import DesignSpace, VariableCollection
 from experiment_design.variable.variable import (
     ContinuousVariable,
-    DesignSpace,
     DiscreteVariable,
+    Variable,
     create_continuous_uniform_variables,
     create_discrete_uniform_variables,
     create_distribution_representation,
