+++ tests/test_variable.py
@@ -3,6 +3,7 @@ import pytest
 from scipy import stats
 
 import experiment_design.variable as module_under_test
+import experiment_design.variable.space
 
 
 @pytest.fixture
@@ -26,9 +27,9 @@ def standard_normal() -> module_under_test.ContinuousVariable:
     ),
     ids=("Continuous", "Mixed", "Discrete"),
 )
-def design_space(request) -> module_under_test.DesignSpace:
+def design_space(request) -> experiment_design.variable.space.DesignSpace:
     variables = module_under_test.create_variables_from_distributions(request.param)
-    return module_under_test.DesignSpace(variables)
+    return experiment_design.variable.space.DesignSpace(variables)
 
 
 def test_is_frozen_discrete():
@@ -185,7 +186,7 @@ class TestDiscreteVariable:
 
 class TestDesignSpace:
 
-    def test_value_of(self, design_space: module_under_test.DesignSpace):
+    def test_value_of(self, design_space: experiment_design.variable.space.DesignSpace):
         probabilities = np.array([[0.1, 0.1], [0.5, 0.6], [0.9, 0.8]])
         if isinstance(design_space.variables[0], module_under_test.DiscreteVariable):
             # Both are discrete
@@ -199,13 +200,13 @@ class TestDesignSpace:
         assert np.all(np.isclose(design_space.value_of(probabilities), expected))
 
     def test_design_space_size_getters(
-        self, design_space: module_under_test.DesignSpace
+        self, design_space: experiment_design.variable.space.DesignSpace
     ):
         assert design_space.dimensions == 2
         assert len(design_space) == 2
 
     def test_design_space_lower_bound(
-        self, design_space: module_under_test.DesignSpace
+        self, design_space: experiment_design.variable.space.DesignSpace
     ):
         if isinstance(design_space.variables[0], module_under_test.DiscreteVariable):
             # Both are discrete
@@ -219,7 +220,7 @@ class TestDesignSpace:
         assert all(design_space.lower_bound == expected)
 
     def test_design_space_upper_bound(
-        self, design_space: module_under_test.DesignSpace
+        self, design_space: experiment_design.variable.space.DesignSpace
     ):
         if isinstance(design_space.variables[0], module_under_test.DiscreteVariable):
             # Both are discrete
